requirements.txt
flask
requests
websocket-client
resend
